# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "infinitas hosted registry",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/skills/{skill_id}/data-snapshots": {
      "post": {
        "tags": [
          "authoring-data-snapshots"
        ],
        "summary": "Register Snapshot",
        "operationId": "register_snapshot_api_v1_skills__skill_id__data_snapshots_post",
        "parameters": [
          {
            "name": "skill_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Skill Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "infinitas_auth_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Infinitas Auth Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataSnapshotRegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataSnapshotView"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Skill, version, or data snapshot not found"
          },
          "409": {
            "description": "Data snapshot conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ]
      },
      "get": {
        "tags": [
          "authoring-data-snapshots"
        ],
        "summary": "List Snapshots",
        "operationId": "list_snapshots_api_v1_skills__skill_id__data_snapshots_get",
        "parameters": [
          {
            "name": "skill_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Skill Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "infinitas_auth_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Infinitas Auth Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DataSnapshotView"
                  },
                  "title": "Response List Snapshots Api V1 Skills  Skill Id  Data Snapshots Get"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Skill, version, or data snapshot not found"
          },
          "409": {
            "description": "Data snapshot conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ]
      }
    },
    "/api/v1/skills/{skill_id}/data-snapshots/{snapshot_id}": {
      "get": {
        "tags": [
          "authoring-data-snapshots"
        ],
        "summary": "Get Snapshot",
        "operationId": "get_snapshot_api_v1_skills__skill_id__data_snapshots__snapshot_id__get",
        "parameters": [
          {
            "name": "skill_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Skill Id"
            }
          },
          {
            "name": "snapshot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^dsp_[A-Za-z0-9_-]+$",
              "title": "Snapshot Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "infinitas_auth_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Infinitas Auth Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataSnapshotView"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Skill, version, or data snapshot not found"
          },
          "409": {
            "description": "Data snapshot conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DataSnapshotRegisterRequest": {
        "properties": {
          "skill_version_id": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Skill Version Id"
          },
          "parent_snapshot_id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^dsp_[A-Za-z0-9_-]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Snapshot Id"
          },
          "schema_version": {
            "type": "integer",
            "maximum": 1000000.0,
            "minimum": 1.0,
            "title": "Schema Version"
          },
          "encrypted_object_uri": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 8,
            "title": "Encrypted Object Uri"
          },
          "ciphertext_sha256": {
            "type": "string",
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Ciphertext Sha256"
          },
          "ciphertext_size_bytes": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Ciphertext Size Bytes"
          },
          "manifest_digest": {
            "type": "string",
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Manifest Digest"
          },
          "encryption": {
            "type": "string",
            "const": "age",
            "title": "Encryption",
            "default": "age"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "skill_version_id",
          "schema_version",
          "encrypted_object_uri",
          "ciphertext_sha256",
          "ciphertext_size_bytes",
          "manifest_digest"
        ],
        "title": "DataSnapshotRegisterRequest"
      },
      "DataSnapshotView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "skill_id": {
            "type": "integer",
            "title": "Skill Id"
          },
          "skill_version_id": {
            "type": "integer",
            "title": "Skill Version Id"
          },
          "parent_snapshot_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Snapshot Id"
          },
          "schema_version": {
            "type": "integer",
            "title": "Schema Version"
          },
          "encrypted_object_uri": {
            "type": "string",
            "title": "Encrypted Object Uri"
          },
          "ciphertext_sha256": {
            "type": "string",
            "title": "Ciphertext Sha256"
          },
          "ciphertext_size_bytes": {
            "type": "integer",
            "title": "Ciphertext Size Bytes"
          },
          "manifest_digest": {
            "type": "string",
            "title": "Manifest Digest"
          },
          "encryption": {
            "type": "string",
            "const": "age",
            "title": "Encryption"
          },
          "state": {
            "type": "string",
            "enum": [
              "registered",
              "retired"
            ],
            "title": "State"
          },
          "created_by_principal_id": {
            "type": "integer",
            "title": "Created By Principal Id"
          },
          "actor": {
            "additionalProperties": true,
            "type": "object",
            "title": "Actor"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "skill_id",
          "skill_version_id",
          "parent_snapshot_id",
          "schema_version",
          "encrypted_object_uri",
          "ciphertext_sha256",
          "ciphertext_size_bytes",
          "manifest_digest",
          "encryption",
          "state",
          "created_by_principal_id",
          "created_at"
        ],
        "title": "DataSnapshotView"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "infinitas_auth_token"
      }
    }
  }
}